[Mission2] JOLEE - 보너스 구현사항: Event delegate #81

+++ mission2/8th/jolee/src/js/TodoList.js
@@ -8,29 +8,26 @@ function TodoList(toggleTodo, deleteTodo) {
 
   const $todoList = document.querySelector('#todo-list')
 
+  $todoList.addEventListener('click', (event) => {
+    event.stopPropagation()
+    if (event.target.className === 'todo-item') {
+      const index = Number(event.target.getAttribute('key'))
+      this.toggleTodo(index)
+    } else if (event.target.className === 'delete-button') {
+      const index = Number(event.target.parentNode.getAttribute('key'))
+      this.deleteTodo(index)
+    }
+  })
+
   this.render = function () {
     $todoList.innerHTML = this.data
       .map(
-        (todo) =>
-          `<li class="todo-item">${
+        (todo, index) =>
+          `<li key="${index}" class="todo-item">${
             todo.isCompleted ? `<s>${todo.text}</s>` : `${todo.text}`
           } <button class="delete-button">삭제</button></li>`
       )
       .join('')
-
-    const $todoItems = document.querySelectorAll('.todo-item')
-    $todoItems.forEach(($todoItem, index) => {
-      $todoItem.addEventListener('click', () => {
-        this.toggleTodo(index)
-      })
-    })
-
-    const $deleteTodoBtns = document.querySelectorAll('.delete-button')
-    $deleteTodoBtns.forEach(($deleteTodoBtn, index) => {
-      $deleteTodoBtn.addEventListener('click', () => {
-        this.deleteTodo(index)
-      })
-    })
 
     this.todoCount = new TodoCount(this.data)
   }
